+++ isy-security/src/main/java/de/bund/bva/isyfact/security/xmlparser/RolePrivilegesMapper.java
@@ -4,8 +4,9 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.util.Collection;
 import java.util.HashMap;
-import java.util.List;
+import java.util.HashSet;
 import java.util.Map;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 import com.fasterxml.jackson.dataformat.xml.XmlMapper;
@@ -18,44 +19,43 @@ public class RolePrivilegesMapper {
     /** Logger. */
     private static final IsyLogger LOG = IsyLoggerFactory.getLogger(RolePrivilegesMapper.class);
 
-    private final RolePrivileges rolePrivileges;
+    private final String applicationId;
+
+    private final Map<String, Set<String>> rolePrivilegesMap;
 
     public RolePrivilegesMapper(String roleMappingXmlFilePath) {
-        rolePrivileges = getRolePrivileges(roleMappingXmlFilePath);
-    }
-
-    public List<String> getPrivilegesByRoles(Collection<String> roles) {
-        return roles.stream()
-                .map(this::getPrivilegesByRole)
-                .flatMap(Collection::stream)
-                .collect(Collectors.toList());
-    }
-
-    public List<String> getAllPrivileges() {
-        return rolePrivileges.getPrivileges().stream()
-                .map(Privilege::getId)
-                .collect(Collectors.toList());
-    }
-
-    public Map<String, List<String>> getRolePrivilegesMap() {
-        Map<String, List<String>> result = new HashMap<>();
-        for (Role r : rolePrivileges.getRoles()) {
-            List<String> privilegesForRole = r.getPrivileges().stream()
-                    .map(Privilege::getId)
-                    .collect(Collectors.toList());
-            result.put(r.getId(), privilegesForRole);
+        RolePrivileges rolePrivileges = getRolePrivileges(roleMappingXmlFilePath);
+        rolePrivilegesMap = mapAndValidateRolePrivileges(rolePrivileges);
+        applicationId = rolePrivileges.getApplicationId();
+        if (applicationId == null || applicationId.isEmpty()) {
+            throw new RolePrivilegesMappingException("Es ist keine AnwendungsId gesetzt");
         }
-        return result;
     }
 
-    public String getAnwendungsId() {
-        return rolePrivileges.getAppId();
+    public Set<String> getPrivilegesByRoles(Collection<String> roles) {
+        return roles.stream()
+                .map(rolePrivilegesMap::get)
+                .flatMap(Set::stream)
+                .collect(Collectors.toSet());
+    }
+
+    public Set<String> getAllPrivileges() {
+        return rolePrivilegesMap.values().stream()
+                .flatMap(Set::stream)
+                .collect(Collectors.toSet());
+    }
+
+    public Map<String, Set<String>> getRolePrivilegesMap() {
+        return rolePrivilegesMap;
+    }
+
+    public String getApplicationId() {
+        return applicationId;
     }
 
     @Override
     public String toString() {
-        return "AnwendungsId: " + rolePrivileges.getAppId() + "\n" +
-                "RollenRechteMapping: " + getRolePrivilegesMap().toString();
+        return "AnwendungsId: " + applicationId + "\nRollenRechteMapping: " + getRolePrivilegesMap().toString();
     }
 
     private RolePrivileges getRolePrivileges(String roleMappingXmlFilePath) {
@@ -64,18 +64,42 @@ public class RolePrivilegesMapper {
             InputStream roleMappingFile = getClass().getResourceAsStream(roleMappingXmlFilePath);
             return mapper.readValue(roleMappingFile, RolePrivileges.class);
         } catch (IOException e) {
-            LOG.error(null, "Error loading XML file", e);
-            return null;
+            throw new RolePrivilegesMappingException("Error loading XML file: " + roleMappingXmlFilePath, e);
         }
     }
 
-    private List<String> getPrivilegesByRole(String userRole) {
-        return rolePrivileges.getRoles().stream()
-                .filter(role -> role.getId().equals(userRole))
-                .map(Role::getPrivileges)
-                .flatMap(Collection::stream)
-                .map(Privilege::getId)
-                .collect(Collectors.toList());
+    private Map<String, Set<String>> mapAndValidateRolePrivileges(RolePrivileges rolePrivileges) {
+        Set<String> definedPrivileges = rolePrivileges.getPrivileges().stream().map(privilege1 -> {
+            String privilegeId = privilege1.getId();
+            if (privilegeId == null || privilegeId.isEmpty()) {
+                throw new RolePrivilegesMappingException("Recht ohne RechtId gefunden");
+            }
+            return privilegeId;
+        }).collect(Collectors.toSet());
+
+        Map<String, Set<String>> result = new HashMap<>();
+        for (Role role : rolePrivileges.getRoles()) {
+            String roleId = role.getId();
+            if (roleId == null || roleId.isEmpty()) {
+                throw new RolePrivilegesMappingException("ID der Rolle fehlt");
+            }
+            Set<String> privilegesForRole = new HashSet<>();
+            for (Privilege privilege : role.getPrivileges()) {
+                String privilegeId = privilege.getId();
+                if (privilegeId == null || privilegeId.isEmpty()) {
+                    throw new RolePrivilegesMappingException("Recht hat keine ID");
+                }
+                if (!definedPrivileges.contains(privilegeId)) {
+                    throw new RolePrivilegesMappingException("Das Recht mit der ID \"" + privilegeId + "\" ist unbekannt");
+                }
+                privilegesForRole.add(privilegeId);
+            }
+            result.put(roleId, privilegesForRole);
+        }
+        if (result.keySet().isEmpty()) {
+            throw new RolePrivilegesMappingException("Es wurde keine einzige Rolle gefunden");
+        }
+        return result;
     }
 
 }